Updates README and moves json parser to utils folder
