+++ server/src/services/webScraping/crawlers/sportsCrawler.py
@@ -12,6 +12,7 @@ import os
 
 from ..freshRSS.sportsLinks import extract_yahoo_sports_links
 from ..models.sportsModel import SportsArticleExtraction
+from ..utils.jsonOutputParser import save_json_pretty
 
 # Get fresh links from RSS feed
 sports_links = extract_yahoo_sports_links()
@@ -34,55 +35,6 @@ MAX_TOKEN = 100000 # this can be adjusted
 
 # Initialize the FirecrawlApp with your API key
 app = FirecrawlApp()
-
-# Function to save JSON array to a file in a pretty-printed format
-def save_json_pretty(data, filename):
-    """
-    Append data to an existing JSON file or create a new one if it doesn't exist.
-    
-    Args:
-        data: The new data to append (can be Pydantic model or dict)
-        filename (str): The name of the file
-    """
-    try:
-        # Set path to outputs directory
-        filepath = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'outputs', filename)
-        
-        # Convert Pydantic model to dict if necessary
-        if hasattr(data, 'model_dump'):  # For Pydantic v2
-            data = data.model_dump()
-        elif hasattr(data, 'dict'):      # For Pydantic v1
-            data = data.dict()
-            
-        # Initialize existing_data as an empty list
-        existing_data = []
-        
-        # Try to read existing file
-        try:
-            with open(filepath, "r", encoding="utf-8") as file:
-                existing_data = json.load(file)
-                if not isinstance(existing_data, list):
-                    existing_data = [existing_data]
-        except FileNotFoundError:
-            print(f"Creating new file: {filepath}")
-        except json.JSONDecodeError:
-            print(f"Error reading existing file. Creating new file: {filepath}")
-            
-        # Append new data
-        if isinstance(data, list):
-            existing_data.extend(data)
-        else:
-            existing_data.append(data)
-            
-        # Write back to file
-        with open(filepath, "w", encoding="utf-8") as file:
-            json.dump(existing_data, file, indent=4, sort_keys=True, ensure_ascii=False)
-        print(f"Data successfully appended to {filepath}")
-        
-    except Exception as e:
-        print(f"An error occurred: {e}")
-
-
 
 # Start timer for first loop
 start_time_crawl = time.time()
