+++ server/src/services/webScraping/utils/__init__.py
@@ -1,0 +1,1 @@
+# Empty init file to make this a package
